Add Logging Support
At this point, calling chromaprint_feed will occasionally throw a
corrupted memory exception.  This far, all attempts to rectify have
failed.  Wondering if it is due to the native code acting on the
memory.  So starting a different approach of creating a single wrapper
function, that will contain all the chromaprint steps, in a native
module.